Restructure all ops code into cmd package
Currently ops code is becoming bigger, as we are adding support for
more commands. Breaking it down so that we maintain modularity.
