+++ cmd/root.go
@@ -1,0 +1,20 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func GetRootCommand() *cobra.Command {
+	var rootCmd = &cobra.Command{Use: "ops"}
+	rootCmd.AddCommand(RunCommand())
+	rootCmd.AddCommand(NetCommands())
+	rootCmd.AddCommand(BuildCommand())
+	rootCmd.AddCommand(ManifestCommand())
+	rootCmd.AddCommand(VersionCommand())
+	rootCmd.AddCommand(UpdateCommand())
+	rootCmd.AddCommand(PackageCommands())
+	rootCmd.AddCommand(LoadCommand())
+	rootCmd.AddCommand(InstanceCommands())
+	rootCmd.AddCommand(ImageCommands())
+	return rootCmd
+}
